+++ examples/issue21.js
@@ -1,3 +1,6 @@
+/*
+ * Next query executed after failed query returns invalid output #21
+ */
 const basex = require("../index");
 var session = new basex.Session()
 basex.debug_mode = false;