+++ examples/manysessions.js
@@ -1,10 +1,10 @@
 // start multiple sessions
-// tested with 2000 sessions apr 2014 node 0.10.26, basex 7.8.2
+// tested with 1000 sessions mar 2017 node 4.4.7, basex 8.6.1
 // was failing with around 200 sessions
 var basex =require("../index");
 var log = require("../debug");
 basex.debug_mode = false;
-var sCount=300; // max 200
+var sCount=1000; // max 200
 var sessions=[];
 //show supplied msg then basex server response
 /**